pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "databases>=0.9.0",
    "fastapi",
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
    "python-dotenv==1.1.1",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.5.0",
]
